Рефакторинг
- Вся логика перенесена в сервисы
- Добавлены мапперы для всех сущностей
- Мелкие исправления
