--- src/main/java/ru/geekbrains/webstore/entities/Product.java
@@ -1,55 +1,0 @@
-package ru.geekbrains.webstore.entities;
-
-import java.util.List;
-import java.util.Objects;
-import javax.persistence.CascadeType;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.experimental.Accessors;
-
-@Data
-@Entity
-@Table(name = "products")
-@NoArgsConstructor
-@Accessors(chain = true)
-public class Product {
-
-  @Id
-  @Column(name = "id")
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  @Column(name = "title")
-  private String title;
-
-  @Column(name = "price")
-  private Long price;
-
-  @OneToMany(mappedBy = "product", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
-  private List<Order> orders;
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    Product product = (Product) o;
-    return id.equals(product.id) && title.equals(product.title) && price.equals(product.price);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(id, title, price);
-  }
-}
