--- src/main/java/ru/geekbrains/webstore/entities/Role.java
@@ -1,27 +1,0 @@
-package ru.geekbrains.webstore.entities;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.experimental.Accessors;
-
-@Entity
-@Data
-@Table(name = "roles")
-@NoArgsConstructor
-@Accessors(chain = true)
-public class Role {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "id")
-  private Long id;
-
-  @Column(name = "name")
-  private String name;
-}
