+++ src/main/java/ru/geekbrains/webstore/exception/DataValidationException.java
@@ -1,0 +1,15 @@
+package ru.geekbrains.webstore.exception;
+
+import java.util.List;
+import lombok.Getter;
+
+public class DataValidationException extends RuntimeException {
+
+  @Getter
+  private final List<String> messages;
+
+  public DataValidationException(List<String> messages) {
+    this.messages = messages;
+  }
+
+}
