+++ src/main/java/ru/geekbrains/webstore/exception/ResourceNotFoundException.java
@@ -1,0 +1,9 @@
+package ru.geekbrains.webstore.exception;
+
+public class ResourceNotFoundException extends RuntimeException {
+
+  public ResourceNotFoundException(String message) {
+    super(message);
+  }
+
+}
