--- src/main/java/ru/geekbrains/webstore/exceptions/GlobalExceptionHandler.java
@@ -1,20 +1,0 @@
-package ru.geekbrains.webstore.exceptions;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@ControllerAdvice
-public class GlobalExceptionHandler {
-
-  @ExceptionHandler
-  public ResponseEntity<?> catchResourceNotFoundException(ResourceNotFoundException e) {
-    return new ResponseEntity<>(new ErrorResponse(e.getMessage()), HttpStatus.NOT_FOUND);
-  }
-
-  @ExceptionHandler
-  public ResponseEntity<?> catchDataValidationException(DataValidationException e) {
-    return new ResponseEntity<>(new ErrorResponse(e.getMessages()), HttpStatus.BAD_REQUEST);
-  }
-}
