--- src/main/java/ru/geekbrains/webstore/exceptions/ResourceNotFoundException.java
@@ -1,9 +1,0 @@
-package ru.geekbrains.webstore.exceptions;
-
-public class ResourceNotFoundException extends RuntimeException {
-
-  public ResourceNotFoundException(String message) {
-    super(message);
-  }
-
-}
