--- src/main/java/ru/geekbrains/webstore/repositories/ProductRepository.java
@@ -1,19 +1,0 @@
-package ru.geekbrains.webstore.repositories;
-
-import java.util.List;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import ru.geekbrains.webstore.entities.Product;
-
-@Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
-
-  List<Product> findAllByPriceIsLessThanEqual(Double maxPrice);
-
-  List<Product> findAllByPriceGreaterThanEqual(Double minPrice);
-
-  List<Product> findAllByPriceBetween(Double minPrice, Double maxPrice);
-
-  Optional<Product> findProductByTitle(String title);
-}
