--- src/main/java/ru/geekbrains/webstore/repositories/UserRepository.java
@@ -1,12 +1,0 @@
-package ru.geekbrains.webstore.repositories;
-
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import ru.geekbrains.webstore.entities.User;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-  Optional<User> findUserByUsername(String name);
-}
