+++ src/main/java/ru/geekbrains/webstore/repository/OrderRepository.java
@@ -1,0 +1,10 @@
+package ru.geekbrains.webstore.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import ru.geekbrains.webstore.entity.Order;
+
+@Repository
+public interface OrderRepository extends JpaRepository<Order, Long> {
+
+}
