+++ src/main/java/ru/geekbrains/webstore/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package ru.geekbrains.webstore.repository;
+
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import ru.geekbrains.webstore.entity.User;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+  Optional<User> findUserByUsername(String name);
+}
