+++ src/main/resources/static/orders.html
@@ -30,7 +30,7 @@
     <tr ng-repeat="o in ordersPage.content">
       <td>{{o.id}}</td>
       <td>{{o.purchasePrise}}</td>
-      <td>{{o.userName}}</td>
+      <td>{{o.username}}</td>
       <td>{{o.productTitle}}</td>
       <td>
         <button class="btn btn-primary" ng-click="deleteOrder(o)">Delete</button>
